requirements.txt
fastapi==0.115.6
pydantic_settings==2.7.0
jwt==1.3.1
passlib==1.7.4